mem-cache: power-off nonvalid blks functionallity added.
The non-valid blocks in the begging of a simulation were
not got decayed. Now we let them decay as this is a more
proper way for the decay mechanism to function.

Also we added a small but I think important modification
in decay mechanism (useful in IATAC) to call handleHit
on a regular hit too.

Change-Id: I0b853a5727c59bb3054ea9377bb0b2813dabe54a