mem-cache: various fixes on tour decay
Number of dedicated sets is set to 32. Now it is more
clear how to set the number you want (e.g. 16, 64). We
want in the future to make it a modifiable parameter.

Tournament window is now fixed around 4M. Also, we added an
upper limit on global decay counter so to avoid being the
global counter bigger than the window.

Lastly, a new tournament decay variation is implemented. Now
we consider std leader team misses on top of decay-induced misses
in order to decide to downscale/upscale

Change-Id: I4a41a409a98c17d04352c33b37c856b9c8064e72