mem-cache: functionality for amc decay mechanism support added.
A new amc event handler added in order for the amc
to work.

Inside the cache, a new gen_event_handler is added.
Currently it is used for the amc, but with little
modifications it is going to be used for any type
of decay. In order for this to be possible, we further
added an enum to characterize a TimingEventHandler as
IATACEventHandler, DecayEventHandler and so on.

Change-Id: Ie500391e35d5fd9a32700404a4d6b046c9bbc3b6