mem-cache: di/2 role on j-scale got canceled.
We used to count di/2 DIMs also in order to decide
whether to do a jump upscale. This was not completely
right, especially on cases where di was pretty good.

Also, we modified a check on tournament decay, so for the
case of (di, 2di) = (0, 0), the mechanism should now stay
at the current di (and not upscale, as before)

Change-Id: I8ac42e83493760f7d60ef9e197a64fd88fd46074