mem-cache: Update on tour-decay: parameterization and fixes.
Now, it is possible to pass through python config file
the type of tournament variation (plain, jump, ejump, opt
and opt_s).

Also, bug on sample() that could possibly perform a
window-skip on AMC decay got fixed.

We now keep count of Leader Team misses of all LTs and
other tourname stats also (newly decayd blocks per W, total DIMs,
turnoff-% until current W).

Bug that counted write misses as DIMs is now fixed (I hope).

We introduced Tour-UD-S (up & down smoothing), defined as OPT_S in
the code. In this variation, we added a looser, yet more suitable limit
for upscale and downscale adaptive to the LS-to-S ratio (that is, it's
adaptive to cache size and the number of LS).

Change-Id: I09d2b9f5a6898a3fd4d2941a3e703dc62b710670