mem-cache: New Tour-Energy-Aware variation added.
We introduce a new tournament decay variation.
In this variation, we exploit the information
of energy benefit among the leader teams.

If the energy benefit from d/2 decay is higher
than that of d LT, we downscale. If the energy benefit
from 2d decay is higher than that of d LT, we upscale.
If all DIs cause energy increase, we jump-scale.

Energy Benefit is calculated as leakage_energy_saved - dynamic_energy_increase_due_to_misses.

Energy values are calculated for an 8Gb DDR3 and an L3 LLC. These
may not stand correct for a different configuration or for
decay on a higher level cache.

Change-Id: I651f8a85f3f14dba2a00db490ce36826c140b976