mem-cache: masks for stuck-1s and stuck-0s bits added on blks.
We want to add functionality for cache block (lines)
where some bits are stuck in a specifiv value (0 or 1).
The stuck bits are represented with two masks for every byte
of a cache block:
1) maskOnes: 00001000, with 1s in bits stuck in value 1 and 0 everywhere else.
2) maskZeros: 11111011, with 0s in bits stuck in value 0 and 1 everywhere else.

In order to extract the value of a (byte of the) block inserted in a
cache line with stuck bits, you can apply the masks in the following way:
modified_data[i] = (data[i] | maskOnes[i]) & maskZeros[i]

The modification of the data has not been implemented yet.

Change-Id: I211e296d4bce46a1e874a59e353a919c08301669