mem-cache:cache-blk stuck-bits functionality added.
The modification of a faulty-cache's data is now available.
The way they are modified based on a set of faulty/stuck bits
is explained in the previous commit.

Also, base functionality where cache blocks that are noted as
faulty are disabled and skipped by the (re)placement policy
is still available.

Change-Id: I5b1b9af2a7633b0a7fbcab967e5ebf9753ff024b