mem-cache: added functionality for set sampling policies
We added the base code for set sampling policies
(SSPs) in caches.
The policies that are currently available are
NoSSP (all sets are samples) and UniformSSP
(one set every K sets is a sample).

The default policy is NoSSP, in order not to
affect the normal gem5 behavior.

This policies can be used for example to test functionality
that uses counters in order to save energy by using
only the counters of some sets (the sample ones).

Change-Id: Ib144b2a3eb0fe6fad69e57a441338cf8c9cd9537