mem-cache: added simpler way to set sample in SSP
We define a simple way to get the 1st of every N sets
as a sample set.

We find it easy to elicit if a block is going to be in the
sample set based on its set index.

This new way of marking sample sets (and blocks) is not
completed.

This is an indermediate commit for practical reasons.

Change-Id: I50e758bfddcecc1d51618ded899e45dd2d162c93