ext: gem5-nvdla-runtime from c to cc
I'm trying to turn all gem5-nvdla-runtime
into cc (c++) code to be more consistent
and easy to communicate info between files.

I think the basic reason for using c was
that kmd was ment to be part of linux kernel.

That is not needed in our ported code so
c language dependency could potentially be
eliminated.

Change-Id: I9b0bddd6caea382c85ae9bc35243886635f06398