dev: fs_bigLITTLE_RTL.py mod for custom nvdla_runtime
We try to modify nvdla_runtime to run in gem5 arm FS.
For this reason, we modify fs_bigLITTLE_RTL.py to add
a memory for dla dma operations and an IsaFake device
for simulating dla device instance.

These two objects are already hardcodedly defined in the
device tree we pass to fs simulation in order for
nvdla_runtime to work.

Change-Id: Ib9569d869da91dd7e5aa34fd418504dbecfff3bc