dev: [temp commit] NvDlaDevice to BasicPioDevice
Trying to turn NvDlaDevice to BasicPioDevice
(from rtlObject). I put all rtlObject code in
NvDlaDevice to remove any rtlObject dependency.

Next step is to clean any reduntant code that had
to do with nvdla trace (we are going to read/write to
dla through the system - io bus, so the predefined trace
is no longer needed).

Change-Id: Ia4979791623aa4d1fd56273d73d380ac9848608a